requirements.txt
Django==4.1.5
django-filter==23.1
django-allauth==0.65.3
